requirements.txt
fastapi
uvicorn
pydantic
pyodbc==4.0.27
python-dotenv
python-keycloak
pyjwt